+++ lib/vroxy/docs.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+require "vroxy/docs/definition"
+require "vroxy/docs/registry"
+require "vroxy/docs/sync"